myapp/backend/requirements.txt
Django
djangorestframework
psycopg2-binary
sentence-transformers
faiss-cpu
python-dotenv